#146 / SplashActivity 삭제 및 IntroActivity 생성, 로직 이동
