+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.SplashActivity"
+            android:name=".presentation.IntroActivity"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
